add new features2

+++ src/main/kotlin/com/alejandro/kotlin/business/company/CompanyBusinessImpl.kt
@@ -6,11 +6,13 @@ import com.alejandro.kotlin.dto.CompanyDto
 import com.alejandro.kotlin.dto.WebSiteDto
 import com.alejandro.kotlin.entity.CompanyEntity
 import com.alejandro.kotlin.repository.CompanyRepository
+import com.alejandro.kotlin.repository.WebSiteRepository
 import com.alejandro.kotlin.util.FileUtil
+import com.alejandro.kotlin.util.constants.MyConstants
+import com.alejandro.kotlin.util.exception.DuplicatedNameException
+import com.alejandro.kotlin.util.exception.InvalidLogoException
 import com.alejandro.kotlin.util.normalize.Normalizer
 import org.slf4j.LoggerFactory
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.core.io.Resource
 import org.springframework.data.domain.*
 import org.springframework.stereotype.Service
@@ -20,17 +22,18 @@ import java.util.stream.Collectors
 
 @Service
 @Transactional
-class CompanyBusinessImpl @Autowired constructor(
+class CompanyBusinessImpl constructor(
     val companyRepository: CompanyRepository,
-    @Qualifier(value = "img")
-    val fileComponent: FileComponent): CompanyBusiness {
+    val webSiteRepository: WebSiteRepository,
+    val fileComponent: FileComponent
+) : CompanyBusiness {
 
     private val logger = LoggerFactory.getLogger(SpringKotlinExampleApplication::class.java)
 
     override fun findById(id: Long): CompanyDto {
-        if(this.companyRepository.existsById(id)) {
+        if (this.companyRepository.existsById(id)) {
             val response = this.companyRepository.getById(id)
-            logger.info("Find by id: {}",response.toString())
+            logger.info("Find by id: {}", response.toString())
             return response.toDto()
         }
         throw  NoSuchElementException(super.TYPE_ELEMENT + " not found")
@@ -50,11 +53,13 @@ class CompanyBusinessImpl @Autowired constructor(
         throw  NoSuchElementException(super.TYPE_ELEMENT + " not found")
     }
 
-    override fun create(element: CompanyDto): CompanyDto {
+    override fun create(element: CompanyDto): String {
+        this.validNames(element)
         val entity: CompanyEntity = element.toEntity()
         entity.updateWebSites()
         logger.info("Created: {}", entity.toString())
-        return this.companyRepository.save(entity).toDto()
+        val idGenerated = this.companyRepository.save(entity).id.toString()
+        return "$TYPE_ELEMENT/$idGenerated"
     }
 
     override fun update(id: Long, element: CompanyDto): CompanyDto {
@@ -93,9 +98,7 @@ class CompanyBusinessImpl @Autowired constructor(
     override fun removeWebSites(id: Long, webSites: Collection<WebSiteDto>): CompanyDto {
         if (this.companyRepository.existsById(id)) {
             val toUpdate: CompanyEntity = this.companyRepository.getById(id)
-            webSites.forEach {ws ->
-                toUpdate.removeWebSite(ws.toEntity())
-            }
+            webSites.forEach { toUpdate.removeWebSite(it.toEntity()) }
             toUpdate.updateWebSites()
             return this.companyRepository.save(toUpdate).toDto()
         }
@@ -111,12 +114,19 @@ class CompanyBusinessImpl @Autowired constructor(
     }
 
     override fun uploadLogo(logo: MultipartFile, id: Long): Boolean {
+
+        val spliced = logo.originalFilename?.split(".")
+        val extName = spliced?.get(1)
+        if (!MyConstants.LOGO_EXTENSIONS.contains(extName)) {
+            throw InvalidLogoException(extName!!)
+        }
+
         if (this.companyRepository.existsById(id)) {
             val toUpdate = this.companyRepository.findById(id).get()
-            if(toUpdate.logo.isNotEmpty()) {
+            if (toUpdate.logo.isNotEmpty()) {
                 val file = FileUtil.toFile(toUpdate.logo)
-                if(FileUtil.exist(file)) {
-                   this.fileComponent.delete(toUpdate.logo)
+                if (FileUtil.exist(file)) {
+                    this.fileComponent.delete(toUpdate.logo)
                 }
             }
             logo.let {
@@ -132,4 +142,14 @@ class CompanyBusinessImpl @Autowired constructor(
         return false
     }
 
+    private fun validNames(companyDto: CompanyDto) {
+        if (this.companyRepository.countByName(companyDto.name) > 0) {
+            throw DuplicatedNameException(companyDto.name)
+        }
+        companyDto.webSites.forEach { ws ->
+            if (this.webSiteRepository.countByName(ws.name) > 0) {
+                throw DuplicatedNameException(ws.name)
+            }
+        }
+    }
 }
